Improved metrics
No more "internal args"; all args are actual arguments that can be specified at runtime. The program now prints its relevant args and has several safety features that ensure that args are properly set (e.g., if we're attaching a classifier (args.attach_classifier), then  we're set to use a gan (args.using_gan), and we won't train the linear classifiers on our own human-made cfs. Also prints the classifier weight as it increases. Also the progress files (with loss for gan, discriminator, etc.) include the args in their title, so that if you change the args and run again, the files won't get overwritten.